fix: 修复获取不到二维码参数的问题
