+++ miniapp/src/app.tsx
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import Taro from '@tarojs/taro';
+import Taro, {getCurrentInstance} from '@tarojs/taro';
 import {Provider} from 'react-redux';
 import configStore from './store';
 import './styles/custom-variables.scss';
@@ -74,24 +74,7 @@ class App extends Component {
     });
   }
 
-  onLoad(options) {
-    const scannedURL = decodeURIComponent(options.q);
-    if (!scannedURL) return;
-    const splittedURL = scannedURL.split('/');
-    const id = splittedURL.pop();
-    const type = splittedURL.pop();
-    if (type === 'campaign') {
-      switch (id) {
-        case 'lung-nung-dieng': {
-          // TODO: 词表 ID 确定下来之后替换掉下方的 mockId
-          Taro.reLaunch({
-            url: 'pages/collectionDetail/collectionDetail?id=mockId',
-          });
-          break;
-        }
-      }
-    }
-  }
+  onLoad() {}
 
   componentDidShow() {}
